+++ replicast.php
@@ -9,7 +9,7 @@
  * Plugin Name:       Replicast
  * Plugin URI:        http://log.pt/
  * Description:       Replicate content across WordPress installs via the WP REST API.
- * Version:           1.0.1
+ * Version:           1.0.2
  * Author:            log.OSCON, Lda.
  * Author URI:        http://log.pt/
  * License:           GPL-2.0+
@@ -45,6 +45,6 @@ if ( ! defined( 'WPINC' ) ) {
  * @since    1.0.0
  */
 \add_action( 'plugins_loaded', function () {
-    $plugin = new \Replicast\Plugin( 'replicast', '1.0.1' );
+    $plugin = new \Replicast\Plugin( 'replicast', '1.0.2' );
     $plugin->run();
 } );