+++ CHANGELOG.md
@@ -1,5 +1,17 @@
-# Change Log
+## Changelog ##
 
-## Version 1.0
+### 1.0.2 ###
+* Replication failures are reflected in selected sites
+* Added Site ID to admin notices.
+* Bug fix. 
 
-* Initial release.
+### 1.0.1 ###
+* Bug fix. 
+
+### 1.0.0 ###
+* Initial release.  
+
+## Upgrade Notice ##
+
+### 1.0 ###
+Initial release.